requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
